#1113 Refactor textarea code in svg-canvas-render
